backend/requirements.txt
Django>=5.2,<6
daphne>=4.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.2
channels>=4.0
channels-redis>=4.2
redis>=5.0
Pillow>=10.0
psycopg2-binary>=2.9
drf-spectacular>=0.27
